+++ src/modify/dankmeme.globed2/GlobedSettingsLayer.cpp
@@ -9,19 +9,18 @@ class $nodeModify(MyGlobedSettingsLayer, GlobedSettingsLayer){
     void modify(){
         
         auto DayAndNightSystem = DayAndNightSystem::create();
-        DayAndNightSystem->setID("Events"_spr);
+        DayAndNightSystem->setID("events"_spr);
         this->addChild(DayAndNightSystem, -1);
         
         auto DayAndNightSystemOverlay = DayAndNightSystemOverlay::create();
-        DayAndNightSystemOverlay->setID("ScreenOverlay"_spr);
+        DayAndNightSystemOverlay->setID("screen-overlay"_spr);
         this->addChild(DayAndNightSystemOverlay, 106);
 
-        if (auto bg = this->getChildByID("background")){
-			bg->setZOrder(-2);
-            if (DayAndNightSystem::events > 0.99f){
-				bg->setVisible(false);
-			}
-		}
+        if (DayAndNightSystem::events > 0.99f){
+            if (auto bg = this->getChildByID("background")){
+			    bg->setVisible(false);
+		    }
+        }
 
     }