+++ src/modify/gd/LevelBrowserLayer.cpp
@@ -11,21 +11,19 @@ class $modify(LevelBrowserLayer) {
 		return false;
 
 		auto DayAndNightSystem = DayAndNightSystem::create();
-		DayAndNightSystem->setID("Events"_spr);
+		DayAndNightSystem->setID("events"_spr);
 		this->addChild(DayAndNightSystem, -1);
 
 		auto DayAndNightSystemOverlay = DayAndNightSystemOverlay::create();
-		DayAndNightSystemOverlay->setID("ScreenOverlay"_spr);
+		DayAndNightSystemOverlay->setID("screen-overlay"_spr);
 		this->addChild(DayAndNightSystemOverlay, 104);
 
-		if (auto bg = this->getChildByID("background")){
-			bg->setZOrder(-2);
-			if (DayAndNightSystem::events > 0.99f){
+		if (DayAndNightSystem::events > 0.99f){
+			if (auto bg = this->getChildByID("background")){
 				bg->setVisible(false);
 			}
 		}
 
-	
 		return true;
 	}
 };