+++ src/modify/gd/LevelInfoLayer.cpp
@@ -11,11 +11,11 @@ class $modify(LevelInfoLayer) {
 		return false;
 		
 		auto DayAndNightSystem = DayAndNightSystem::create();
-		DayAndNightSystem->setID("Events"_spr);
+		DayAndNightSystem->setID("events"_spr);
 		this->addChild(DayAndNightSystem, -1);
 
 		auto DayAndNightSystemOverlay = DayAndNightSystemOverlay::create();
-		DayAndNightSystemOverlay->setID("ScreenOverlay"_spr);
+		DayAndNightSystemOverlay->setID("screen-overlay"_spr);
 		this->addChild(DayAndNightSystemOverlay, 106);
 
 		if (DayAndNightSystem::events > 0.99f){
@@ -29,21 +29,21 @@ class $modify(LevelInfoLayer) {
 
 			if (auto normalBar = typeinfo_cast<CCSprite*>(this->getChildByID("normal-mode-bar"))){
 
-                auto cc9fix = NineSlice::create("GJ_progressBar_001.png");
-                cc9fix->setPosition(170, 10);
-				cc9fix->setContentSize({343, 23});
-				cc9fix->setID("CC9Fix"_spr);
-				normalBar->addChild(cc9fix, -1);
+                auto normalProgressBarOutline = NineSlice::create("GJ_progressBar_001.png");
+                normalProgressBarOutline->setPosition(170, 10);
+				normalProgressBarOutline->setContentSize({343, 23});
+				normalProgressBarOutline->setID("normal-progress-bar-outline"_spr);
+				normalBar->addChild(normalProgressBarOutline, -1);
 				
                 normalBar->setOpacity(255);
 			}
             if (auto practiceBar = typeinfo_cast<CCSprite*>(this->getChildByID("practice-mode-bar"))){
 
-                auto cc9fix2 = NineSlice::create("GJ_progressBar_001.png");
-                cc9fix2->setPosition(170, 10);
-				cc9fix2->setContentSize({343, 23});
-				cc9fix2->setID("CC9Fix2"_spr);
-				practiceBar->addChild(cc9fix2, -1);
+                auto practiceProgressBarOutline = NineSlice::create("GJ_progressBar_001.png");
+                practiceProgressBarOutline->setPosition(170, 10);
+				practiceProgressBarOutline->setContentSize({343, 23});
+				practiceProgressBarOutline->setID("practice-progress-bar-outline"_spr);
+				practiceBar->addChild(practiceProgressBarOutline, -1);
 
                 practiceBar->setOpacity(255);
 			}