+++ src/modify/gd/LevelPage.cpp
@@ -10,27 +10,55 @@ class $modify(LevelPage) {
 		if (!LevelPage::init(level))
 		return false;
 
+		if (DayAndNightSystem::events > 0.99f){
+
+			if (auto levelButton = m_levelMenu->getChildByID("level-button")){
+
+				if (auto whiteSprite = levelButton->getChildByID("white-sprite")){
+
+					auto cc9Fix = NineSlice::create("square02b_001.png");
+                	cc9Fix->setPosition(whiteSprite->getPosition());
+					cc9Fix->setContentSize(levelButton->getContentSize());
+					cc9Fix->setID("cc-9-fix-1"_spr);
+					levelButton->addChild(cc9Fix, -1);
+
+				}
+
+			}	
+				
+		}
+
 		if (DayAndNightSystem::events > 2.99f){
 
-			if (auto normalProgressBar = typeinfo_cast<CCSprite*>(this->getChildByID("normal-progress-bar"))){
+			if (auto levelbgButton = typeinfo_cast<CCSpriteBatchNode*>(this->getChildByIDRecursive("background"))){
+				levelbgButton->setVisible(false);
+			}
+			if (auto normalProgressBar = typeinfo_cast<CCSprite*>(this->getChildByIDRecursive("normal-progress-bar"))){
 
-                auto cc9fix2 = NineSlice::create("GJ_progressBar_001.png");
-                cc9fix2->setPosition(170, 10);
-				cc9fix2->setContentSize({343, 23});
-				cc9fix2->setID("CC9Fix2"_spr);
-				normalProgressBar->addChild(cc9fix2, -1);
+                auto normalProgressBarOutline = NineSlice::create("GJ_progressBar_001.png");
+                normalProgressBarOutline->setPosition(170, 10);
+				normalProgressBarOutline->setContentSize({343, 23});
+				normalProgressBarOutline->setID("normal-progress-bar-outline"_spr);
+				normalProgressBar->addChild(normalProgressBarOutline, -1);
 				
                 normalProgressBar->setOpacity(255);
 			}
-            if (auto practiceProgressBar = typeinfo_cast<CCSprite*>(this->getChildByID("practice-progress-bar"))){
+            if (auto practiceProgressBar = typeinfo_cast<CCSprite*>(this->getChildByIDRecursive("practice-progress-bar"))){
 
-                auto cc9fix3 = NineSlice::create("GJ_progressBar_001.png");
-                cc9fix3->setPosition(170, 10);
-				cc9fix3->setContentSize({343, 23});
-				cc9fix3->setID("CC9Fix3"_spr);
-				practiceProgressBar->addChild(cc9fix3, -1);
+                auto practiceProgressBarOutline = NineSlice::create("GJ_progressBar_001.png");
+                practiceProgressBarOutline->setPosition(170, 10);
+				practiceProgressBarOutline->setContentSize({343, 23});
+				practiceProgressBarOutline->setID("practice-progress-bar-outline"_spr);
+				practiceProgressBar->addChild(practiceProgressBarOutline, -1);
 
                 practiceProgressBar->setOpacity(255);
+			}
+
+			if (auto levelProgressDisplay = typeinfo_cast<CCScale9Sprite*>(this->getChildByID("level-progress-display"))){
+				levelProgressDisplay->setOpacity(0);
+			}
+			if (auto levelAchievementsDisplay = typeinfo_cast<CCScale9Sprite*>(this->getChildByID("level-achievements-display"))){
+				levelAchievementsDisplay->setOpacity(0);
 			}
 
 		}
@@ -39,12 +67,34 @@ class $modify(LevelPage) {
 
             if (auto levelButton = m_levelMenu->getChildByID("level-button")){
 
-                auto cc9fix = NineSlice::create("square02b_001.png");
-                cc9fix->setPosition(170, 47.5f);
-				cc9fix->setContentSize(levelButton->getContentSize());
-				cc9fix->setID("CC9Fix"_spr);
-                cc9fix->setColor(ccc3(0, 0, 255));
-				levelButton->addChild(cc9fix, -1);
+				if (auto whiteSprite = levelButton->getChildByID("white-sprite")){
+
+					if (auto levelProgressDisplay = this->getChildByID("level-progress-display")){
+
+						auto cc9Fix2 = NineSlice::create("square02b_001.png");
+                		cc9Fix2->setPosition(130, 35);
+						cc9Fix2->setContentSize(levelProgressDisplay->getContentSize());
+						cc9Fix2->setID("cc-9-fix-2"_spr);
+                		cc9Fix2->setColor(ccc3(0, 0, 140));
+						levelProgressDisplay->addChild(cc9Fix2, -1);
+
+					}
+					if (auto levelAchievementsDisplay = this->getChildByID("level-achievements-display")){
+
+						auto cc9Fix3 = NineSlice::create("square02b_001.png");
+                		cc9Fix3->setPosition(130, 65);
+						cc9Fix3->setContentSize(levelAchievementsDisplay->getContentSize());
+						cc9Fix3->setID("cc-9-fix-3"_spr);
+                		cc9Fix3->setColor(ccc3(0, 0, 140));
+						levelAchievementsDisplay->addChild(cc9Fix3, -1);
+
+					}
+					if (auto getcc9Fix = typeinfo_cast<NineSlice*>(levelButton->getChildByID("ninsam.day_and_night_system/cc-9-fix-1"))){
+						getcc9Fix->setColor(ccc3(0, 0, 140));
+					}
+
+				}
+
             }
 
 		}
@@ -52,16 +102,50 @@ class $modify(LevelPage) {
 
             if (auto levelButton = m_levelMenu->getChildByID("level-button")){
 
-                auto cc9fix = NineSlice::create("square02b_001.png");
-                cc9fix->setPosition(170, 47.5f);
-				cc9fix->setContentSize(levelButton->getContentSize());
-				cc9fix->setID("CC9Fix"_spr);
-                cc9fix->setColor(ccc3(0, 0, 145));
-				levelButton->addChild(cc9fix, -1);
+                if (auto whiteSprite = levelButton->getChildByID("white-sprite")){
+
+					if (auto levelProgressDisplay = (this->getChildByID("level-progress-display"))){
+
+						auto cc9Fix2 = NineSlice::create("square02b_001.png");
+                		cc9Fix2->setPosition(130, 35);
+						cc9Fix2->setContentSize(levelProgressDisplay->getContentSize());
+						cc9Fix2->setID("cc-9-fix-2"_spr);
+                		cc9Fix2->setColor(ccc3(0, 0, 75));
+						levelProgressDisplay->addChild(cc9Fix2, -1);
+						
+					}
+					if (auto levelAchievementsDisplay = this->getChildByID("level-achievements-display")){
+
+						auto cc9Fix3 = NineSlice::create("square02b_001.png");
+                		cc9Fix3->setPosition(130, 65);
+						cc9Fix3->setContentSize(levelAchievementsDisplay->getContentSize());
+						cc9Fix3->setID("cc-9-fix-3"_spr);
+                		cc9Fix3->setColor(ccc3(0, 0, 75));
+						levelAchievementsDisplay->addChild(cc9Fix3, -1);
+						
+					}
+					if (auto getcc9Fix = typeinfo_cast<NineSlice*>(levelButton->getChildByID("ninsam.day_and_night_system/cc-9-fix-1"))){
+						getcc9Fix->setColor(ccc3(0, 0, 75));
+					}
+					
+				}
             }
 			
 		}
 
+		if (auto levelButton = m_levelMenu->getChildByID("level-button")){
+
+			if (auto getcc9Fix = typeinfo_cast<NineSlice*>(levelButton->getChildByID("ninsam.day_and_night_system/cc-9-fix-1"))){
+				if ((DayAndNightSystem::events == 1) || (DayAndNightSystem::events == 2)){
+					getcc9Fix->setVisible(false);
+				}
+				if (auto isOverChargedEnabled = (this->getChildByID("level-progress-display"))){
+					getcc9Fix->setColor(ccc3(0, 0, 0));
+					getcc9Fix->setVisible(true);
+				}
+			}
+
+		}
 
 		return true;
 	}