+++ src/modify/geode.loader/ModsLayer.cpp
@@ -9,25 +9,23 @@ class $nodeModify(MyModsLayer, ModsLayer){
     void modify(){
         
         auto DayAndNightSystem = DayAndNightSystem::create();
-        DayAndNightSystem->setID("Events"_spr);
+        DayAndNightSystem->setID("events"_spr);
         this->addChild(DayAndNightSystem, -2);
         
         auto DayAndNightSystemOverlay = DayAndNightSystemOverlay::create();
-        DayAndNightSystemOverlay->setID("ScreenOverlay"_spr);
+        DayAndNightSystemOverlay->setID("screen-overlay"_spr);
         this->addChild(DayAndNightSystemOverlay, 106);
 
-        if (auto bg = this->getChildByID("SwelvyBG")){
-			bg->setZOrder(-3);
-            if (DayAndNightSystem::events > 0.99f){
-				bg->setVisible(false);
-			}
-		}
-        else if (auto bg2 = this->getChildByID("bg")){
-			bg2->setZOrder(-3);
-			if (DayAndNightSystem::events > 0.99f){
-				bg2->setVisible(false);
-			}
-		}
+        if (DayAndNightSystem::events > 0.99f){
+
+            if (auto bg = this->getChildByID("SwelvyBG")){
+			    bg->setVisible(false);
+		    }
+            else if (auto bg2 = this->getChildByID("bg")){
+			    bg2->setVisible(false);
+		    }
+
+        }
 
     }