+++ src/modify/hiimjustin000.integrated_demonlist/IDListLayer.cpp
@@ -9,25 +9,28 @@ class $nodeModify(MyIDListLayer, IDListLayer){
     void modify(){
 
         auto DayAndNightSystem = DayAndNightSystem::create();
-        DayAndNightSystem->setID("Events"_spr);
+        DayAndNightSystem->setID("events"_spr);
         this->addChild(DayAndNightSystem, -2);
         
         auto DayAndNightSystemOverlay = DayAndNightSystemOverlay::create();
-        DayAndNightSystemOverlay->setID("ScreenOverlay"_spr);
+        DayAndNightSystemOverlay->setID("screen-overlay"_spr);
         this->addChild(DayAndNightSystemOverlay, 106);
 
-        if (auto bg = this->getChildByID("background")){
-			bg->setZOrder(-3);
-            if (DayAndNightSystem::events > 0.99f){
-				bg->setVisible(false);
-			}
-		}
-        if (auto listLayer = typeinfo_cast<GJListLayer*>(this->getChildByID("GJListLayer"))){
-            if (DayAndNightSystem::events > 2.99f){
-				listLayer->setOpacity(255);
+        if (DayAndNightSystem::events > 0.99f){
+            if (auto bg = this->getChildByID("background")){
+			    bg->setVisible(false);
+		    }
+        }
+        if (DayAndNightSystem::events > 2.99f){
+
+            if (auto listLayer = typeinfo_cast<GJListLayer*>(this->getChildByID("GJListLayer"))){
+          
+			    listLayer->setOpacity(255);
                 listLayer->setColor(ccc3(0, 0, 0));
-			}
-		}
+			
+		    }
+
+        }
 		
 
     }