+++ src/modify/limegradient.betterachievements/BetterAchievementLayer.cpp
@@ -9,19 +9,18 @@ class $nodeModify(MyBetterAchievementLayer, BetterAchievementLayer){
     void modify(){
         
         auto DayAndNightSystem = DayAndNightSystem::create();
-        DayAndNightSystem->setID("Events"_spr);
+        DayAndNightSystem->setID("events"_spr);
         this->addChild(DayAndNightSystem, -2);
         
         auto DayAndNightSystemOverlay = DayAndNightSystemOverlay::create();
-        DayAndNightSystemOverlay->setID("ScreenOverlay"_spr);
+        DayAndNightSystemOverlay->setID("screen-overlay"_spr);
         this->addChild(DayAndNightSystemOverlay, 106);
 
-        if (auto bg = this->getChildByID("content-background")){
-			bg->setZOrder(-3);
-            if (DayAndNightSystem::events > 0.99f){
-				bg->setVisible(false);
-			}
-		}
+        if (DayAndNightSystem::events > 0.99f){
+            if (auto bg = this->getChildByID("content-background")){
+			    bg->setVisible(false);
+		    }
+        }
 
     }